requirements.txt
streamlit
tensorflow
keras
Pillow
numpy
streamlit-folium
folium
